libc: Fix strcasecmp_l() and strncasecmp_l() issues.
POSIX.1-2008 specifies that those two functions should be declared by
including <strings.h>, not <string.h> (the latter only has strcoll_l()
and strxfrm_l()):

http://pubs.opengroup.org/onlinepubs/9699919799/functions/strcasecmp.html

It looks like FreeBSD got this wrong in their xlocale stuff originally.

And it looks like the only clean way to properly solve it is to split
<xlocale/_string.h> into <xlocale/_string.h> and <xlocale/_strings.h>,
which this commit does.