top: Fix -Sc option combination (bug #2670)
It was reported that the "top -S" and "top -c" commands work as expected
but "top -Sc" showed repeated lines.  I confirmed a malfunction, but a
different one: all 20 lines showing in console were the same.

The following patch will revert to standard format if a system command is
detected, otherwise the desired "full command" option is shown.

< http://bugs.dragonflybsd.org/issues/2670 >