installer: Fix disk_find() on systems that have both da1 and da10.
In addition to comparing names (up to the length of the disk which
was found previously), also compare name lengths.

While here, remove the disk_find() check in disk_new(). disk_new()
didn't deal with it returning NULL. But because the list of disks
we offer as targets for installing is taken from sysctl kern.disks'
output, it will not have duplicates (else it would likely be a bug
in the kernel).

Reported-by: BSDUX on IRC