installer: Remove a more or less arbitrary check of the swap size.
Except for boxes which have lots of memory combined with a very very
small disk, the installer gives a sane default for swap size by itself
and if the user doesn't change it, all will be good.

However, if the user did choose to adjust the swap size and if it was
to below the main memory size, this check caused it not to be activated
and no dumpdev was written to the rc.conf of the installed system.

Instead of guessing around how much swap is needed to hold a dump,
adding warnings and whatnot, just assume that anyone who tinkers with
the offered defaults is intelligent enough to know possible
consequences and accept their choice in any case.