nrelease: Do not provide default make.conf files.
The git package delivered on the installation media is special in the
sense that it is built with certain options not set to prevent it from
pulling in python and perl as dependencies (to keep the ISOs small,
mainly).

Once the system is installed, storage size is no longer really an issue,
so there is no need to have those options in the make.conf we install.

After this commit, a manually built git package should match the binary
package which we provide.