kernel/acpi: Add acpi_eval_osc() as a general function for running _OSCs.
This is a hopefully true-to-the-spec implementation, as described in
chapter "6.2.11 _OSC (Operating System Capabilities)" of the 5.1 spec.
Handling of the query support flag could be better (i.e., we could
pass the returned buffer back to the caller), but we don't need that
yet.

This also fixes bugs in the the error handling which in our acpi_cpu
specific previous implementation didn't work at all, due to my mistakes.

In-discussion-with: sephe