Move {major,minor} pair into a shifted int type va_rdev in vattr
Rather than having a seperate va_{rmajor,rminor} pair in the
vnode attribute structure in 'sys/vnode.h' move to a unified
shifted integer type 'va_rdev' and use the macros we already have
when major or minor components are needed. A follow up to this
would see the removal of makeudev() and some other functions
that essentially repelicate this functionality elsewhere in the
kernel.

Signed-off-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>