wpa_supplicant: Fix driver list with -h or invalid command switch
If wpa_supplicant was given -h (help) argument to show usage, the NDIS
driver displayed as "(null)".  If wpa_supplicant was given no arguments,
the usage display would also appear, but this time the NDIS driver was
listed.

It seems that the NDIS driver is the only one that must be initialized,
and this initialization was getting skipped on unrecognized commands and
the help command.  The solution is to move the initialization of the
driver to occur before the commands are checked so that the driver data
will always be available.