crt1.o: Fix compilation failure with -DGCRT on clang
Clang has been failing to compile the crt1.c with GCRT macro defined.  The
error message stated the "eprol" label was being defined twice.  Clang was
performing an alternative conditional optimization which duplicated the
assembly including our inserted eprol label.

The workaround is to relocate the eprol label after the handle_static_init
function to get it out of the optimization loop.  The trade-off of this
workaround is that handle_static_init() can no longer be profiled.  The
formerly preceding monstartup belongs to gmon and it controls the range of
PCs the profiling code considers.

Suggested by:	dillon