date(1): Add "-R" flag to use RFC 2822 date and time output format
As stated in the man page, this is equivalent to use
"%a, %d %b %Y %T %z" as the output format while LC_TIME is set to the
"C" locale.

This is compatible with date(1) from the GNU core utilities.

Taken-from: FreeBSD

While here, remove $DragonFly$ line from man page.