OpenSSH: default XAUTH_PATH as a make.conf variable
In the vast majority of use cases, the path to the xauth program will be
/usr/local/bin/xauth, the standard location defined by dports.  However,
it should not be hardcoded becase this is not guaranteed.

This commit modifies OpenSSH makefiles to have an overrideable default
called XAUTH_PATH.  This new variable has been added to the make.conf.5
man page and to /etc/defaults/make.conf.