libm: Sync with FreeBSD (gains 6 long double functions)
The following functions have been implemented:
  - coshl
  - erfcl
  - erfl
  - lgammal
  - sinhl
  - tanhl

Before these were approximated with the double versions using the
"imprecise" macros.  I've left the old ones in place (unlike FreeBSD)
but with symbol versioning so that libraries built with the earlier
versions can link to them.  In other words, there are two versions of
these 6 functions, Df306.1 and DF402.0.