rtld: Allows return pathname in dl_iterate_phdr dlpi_name
This is the behavior of Linux and also of FreeBSD since October 2014.
Until this point, dlpi_name returned the short name of the object unless
it wasn't defined, so the fallback was to return the pathname.

Returning the pathname in all cases is considered more useful and matches
the definition of Linux.

Taken from:	FreeBSD svn 272842 (9 Oct 2014)