rtld: Add a postinit debugger hook
FreeBSD's dtrace(1) uses this hook to halt the "victim" process before its
entry point is called, at which point probes and DOF data are registered
with the kernel.  The r_debg_state hook cannot be used for this purpose,
as it is called before the program's init routines are invoked and in
particular before ODF is registered.  It may or may not be useful for
DragonFly by itself, but if dtrace every comes in, rtld will be ready.

Taken from:	FreeBSD svn 265456 (6 May 2014)
		FreeBSD svn 265578 (7 May 2014)