libm: bring in 16 complex functions from NetBSD
FreeBSD is still missing a number of (obscure) complex functions that
causes GCC conclude that C99 is not supported.  Seven weeks ago, NetBSD
brought the functions into their base.

To avoid a repeat of the unholy mixture of libm from FreeBSD and NetBSD
that we had before, I've created a separate directory for NetBSD functions
called "gaps".  If/when FreeBSD acquires these functions, we can remove
files from "gaps" as we sync their equivalents.  The NetBSD sources were
modified as follows:
  * On the man pages, references to the float and double versions of the
    functions were removed when only the long double version was imported.
    This includes singular/plural changes, and file/function name changes.
  * Changed "#include <math.h>" to "#include "math.h"
  * Removed the include for namespace.h
  * Removed all "#if 0" blocks

The sixteen functions have symbol versions of DF402.0 and are:
  * cpow    cpowf   cpowl
  * clog    clogf   clogl
  * cacosl  casinl  catanl
  * ccoshl  csinhl  ctanhl
  * ccosl   csinl   ctanl
  * cexpl