library versioning: Don't bump major version
Eight libraries have unused yet valid symbol maps.  The original plan was
to activate them together and bump all the major library versions at the
same time.  This never happened because the libc symbol map isn't quite
right yet.  However, it's not necessary to bump the version.  If mapped,
both the versioned and unversioned symbol will be available and will
reference the same location, so no breakage will occur.

Once confirmed, the next commit will remove the "RELEASE36" check.  The
c library guard had changed to RELEASE42, so the intent is to activate
versioning on 7 of the 8 libraries in one step.