library versioning: Activate versioning on 7 libraries
All 4.2 packages currently on the system will continue to work on
a system after this commit.  However, the next bulk run of packages will
build using these symbol versions and will not work on systems older
than this commit.  Part of the joys of running on bleeding edge I think.
The next bulk run won't be for a couple of weeks at least (I'm waiting
on a disruptive and huge patch of perl and pgsql to settle down) so there
is no impending breakage expected.

The newly mapped libraries are: z, ncurses, lzma, edit, archive, md, bz2