kernel: Remove our ancient SCTP support.
It was based on an early stage of FreeBSD's support and never got
beyond that point, due to lack of usage and lack of maintenance.

It is also kind of standing the the way of ongoing work in the
IPv6 stack by sephe.

If anyone ever wants to start again on upgrading and maintaining it,
they can decide then if reverting this commit again makes sense as a
base.

Requested-by: sephe
Approved-by:  dillon