acpi/pstate: Allow missing _PCT and _PSS, as long as _PSD exists
Some BIOSes does not export _PCT and _PSS for the second thread of CPU
cores.  As long as _PSD exists and control/status registers are "fixed
hardware", e.g. for most of the modern Intel CPUs, we could reconstruct
P-state for the CPU even if there are no _PCT or _PSS.

In case that _PSD is wrong, e.g. 2 threads of a same core are in different
P-state domain, we try to put them to the same P-state domain.  This could
be disabled by setting tunable hw.acpi.cpu.pst.ht_reuse_domain to 0; this
tunable is 1 by default.