bsd.hostlib.mk: Fix case where SRCS is defined but empty
This is similar to the recent changes to bsd.lib.mk and bsd.prog.mk.
The check covers two augmentations to OBJS, plus let's add some indents
for readability.  Note the second augmentation didn't have a bug in it,
so including it inside the wrap only saves having to evaluate it.