libc: Silence a bunch of set-but-not-used errors found by gcc 5.0
During testing of upcoming base gcc 5.0, several libc files failed to
compile with -Werror set.  This commit will silence the errors.  For the
majority of these, I designed the variables as "__unused" so the compiler
knows that their (lack of) use is intentional.  For the rest, I just
removed the variables completely.