libstand: Fix loop with undefined optimized behavior
With gcc 4.8+, libstand fails to build with "interation 8u involved in
undefined behavior [-Werror=aggressive-loop-optimizations]".  I figured
out that problem was the code was intentionally defining two fields in
the same structure with a single loop.

The "name" field is an array of size 8 while the "ext" field is an array
of size 3.  The loop in question iterated 11 times to initialize both,
relying on contiguous memory to work.  Apparently this leads to undefined
behavior, so I altered the line to explicitly define the correct element
on the correct field.