libevtr: Fix loop properly
The bruteforce fix should have used elements 0 - 2 because the strings
array was defined as EVTR_NS_MAX - 1, not EVTR_NS_MAX.  So the fix was
wrong, but it exposed the problem with the original code: the upper bound
of the loop was set to EVTR_NS_MAX, not EVTR_NS_MAX - 1 like the "maps"
version below it.

This commit reverts the previous change and fixes the problem with the
original code.

spotted-by:	ivadasz