gcc50: Bring in makefiles but leave them unhooked
The pre-release version of GCC 5.0 will be brought into base soon.  This
commit brings in the makefiles that build it, but they are not hooked
into the build in any way.  Before that happens, the vendor branch needs
to be merged into base, and dozens for files need to be simultaneously
changed.  GCC 4.4 will be unhooked at the same time.

The GCC releases provide pre-generated man pages, but the pre-releases do
not.  To support this, a switch has been added that installs man pages
from these directories.  When GCC 5.0 is released, those man pages will
be removed and the man pages will be installed from contrib/gcc-5.0