Hook gcc 5.0 (pre-release) in build, unhook gcc 4.4
Effectively replace GCC 4.4 with the 1 February 2015 snapshot of the
pre-release GCC 5.0.  This compiler is in pretty good shape, but it is
not perfect.  It cannot fully build world if WORLD_CCCVER is set to
GCC50.  This compiler import is a work in progress.

* The old compiler has been unhooked.
* The gcc44 sources may be removed soon
* Obviously NO_GCC44 doesn't work anymore
* NO_GCC50 has been created and should work
* NO_CXX was never documented and has been removed
* NO_OBJC was not deemed worthwhile and it has been removed
* man pages have been updated accordingly
* default make.conf has been updated accordingly
* Per bsd.sys.mk, gcc50 behaves just like gcc47
* Actually, all CCVER starting with "gcc" now behave like gcc47
* The proper CCVER value is "gcc50".  If "gcc5" is specified, the
  dports compiler lang/gcc5 will be used.

Iterating: THIS IS AN EXPERIMENTAL COMPILER!  It's not ready for
prime-time yet.  Do NOT set WORLD_CCVER to "gcc50".