opiekey (contrib): Fix serious (?) [-Wsizeof-pointer-memaccess]
On a password verification function of opiekey, the amount of bytes
allocated to a secret was 4 bytes of i386 and 8 bytes of x86-64.  Memset
was using the size of the pointer to allocate memory.

Change 3 instances of this to size OPIE_SECRET_MAX + 1 to fix it.
FYI this vendor branch is like 16 years old...