gcc50/csu: Skip depends step to avoid possible race
I've witnessed buildworld break once on muscles at gcc50/csu, on the
depends step.  I can reproduce the error manually by doing:

  > cd gnu/lib/gcc50/csu
  > wmake clean cleandepend
  > wmake depend

It fails to locate insn-constants.h which should be avaiable due to
and include flag.  The depends is just checking if all the headers are
in place.  Skipping the depends step may prevent this breakage.  I have
successfully built world with this patch so it doesn't do harm.