gcc50/stddef.h: Remove __FreeBSD only conditions
In a very special case of DragonFly masquerading as FreeBSD within dports,
using the stock stddef.h that comes with gcc always results in failure.

The masquerading is done by passing __FreeBSD definition through cflags,
and this causes the compiler to attempt to include headers that do not
exist on DragonFly.  The gcc compilers in dports already have this fix.

This commit simply removes all clauses that pertain to FreeBSD and not
DragonFly.