gcc50: Look for ld-elf.so.2 on the root slice always
A couple of us could not could not even get the shell after rebuilding
world with the new compiler.  The problem was that several programs such
as sh, cd, mount*, etc need to run before /usr is mounted when /usr is on
a different slice (those programs are dynamically linked these days)  I
didn't consider this when I decided to leave the interpreter at the default
value which is "/usr/libexec/ld-elf.so.2".  That file is now symlink to
/libexec/ld-elf.so.2.  sh, cd, etc, could not access the dynamic linker
because the mount command failed due to also needing the dynamic linker.

Let's rectify that bad decision now before any other systems get badly
messed up.