gcc50/libstdc++: Use generated symbol map instead
For both gcc47 and gcc50, we've been using pregenerate symbol map for
libstdc++ due to the apparently complexity of generating it.  That's
already come back to bite on gcc50 as several symbols were modified and
added in the last couple of weeks.  It was time to bite the bullet and
figure out how to generate the symbol map.  It turns out that it really
wasn't that hard.

This *should* fix some dports errors (tbc).