make(1): Create custom variable .MAKE.BUILT.BY
The value of .MAKE.BUILT.BY is normally either "gcc47" or "gcc50"
depending on which base compiler built the DragonFly world.  It could
also be any valid CCVER for a dports compiler such as "clang35".

This capability is needed to handle the c++ ABI change that has been
introduced by gcc50.  The problem hits when gcc50 is used to build when
the system libraries have been built by gcc47.  The sys.mk file will
need .MAKE.BUILT.BY to adjust c++ flags appropriately.