sys.mk: Add _GLIBCXX_USE_CXX11_ABI=0 definition in some cases
When the world is built by gcc47, and sys.mk isn't being used to build
the system, the CXXFLAGS will have -D_GLIBCXX_USE_CXX11_ABI=0 added to
it to force gcc50 to use the c++ ABI that gcc47 used when building the
system libraries.

If the world was built with gcc50, the system libraries were built
against the new c++ ABI so this flag would not be added in that case.

Due to new SYSBUILD definition in Makefile.inc1, this flag is never
added during the build of the world.