crunchide(1): fix offset for x86_64, restores WORLD_LDVER
The e_shoff offset for x86_64 is 40, not 44.  FreeBSD figured this out
in January.  The gold linker figured out there was a problem with the
crunch binaries when I tried WORLD_LDVER=ld.gold and it broke on initrd.
The ld.bfd linker didn't squawk.  With the FreeBSD patch, buildworld
completes using the gold linker once again.

While here, tidy up a bit.