kernel/wlan: Fix up the MAC address printing and add back __printflike()s.
In the recent wlan, iwn and ath upgrades, porting this properly was hacked
around by removing the __printflike()s because DragonFly has no support
for %D (neither in GCC nor in kprintf()). This is wrong because even
though it silences the warning, it will still not print MAC addresses
correctly. So bring all that back.

Also update the wlan's README.DRAGONFLY with some information about how
to port this properly.