gcc50: Exclude -Wunused-local-typedefs from C++ flags for now.
At least GCC 5.0 is warning on its own C++ includes (stl_algo.h,
via <algorithm>) if -Wextra is set, breaking the build of programs
which do nothing wrong themselves (such as devd(8)).

When this gets fixed, we can revert this commit again.

Note that I didn't check with GCC 4.8 or 4.9.