Switch default MTA from sendmail to dma
The mailwrapper controls which Mail Transfer Agent is used.  The
mailer.conf defines the behavior of the mailwrapper.  This commit will
have no effect on existing installations, but for the first installation
a new mailer.conf will be installed during the "distribution" target.

This means all new images have the DragonFly Mail Agent acting as base
MTA rather than Sendmail for the first time in the project's history.

While here, set the default path in the mail wrapper to dma as well,
although this isn't used unless the mailer.conf is incomplete.