Effectively remove sendmail from base
The NO_SENDMAIL build knob has been hardwired in sys.mk, so it is not
possible for it to be unset by /etc/make.conf.  This effectively prevents
sendmail from building.

A new check was added to "make upgrade" target to make sure the current
/etc/mail/mailer.conf is not configured to use the base sendmail as the
MTA.  If it fails the check, the upgrade will fail before it starts.
Directions are given hot to fix the problem in the error message.  A
successful check results in the sendmail-related binaries getting removed
from the system.

Finally, NO_SENDMAIL has been removed from the default make.conf and it's
man page.

The actual sendmail makefiles and source remain for now.  Once it is
confirmed that transition is working and dma is functioning well as a
replacement, sendmail and the NO_SENDMAIL conditions will be removed
permanently, along with modifying the mtree files.

Approved-by:	dillon