acpi/cstate: Match C-states of the hyperthreads on the same core
Some stupid BIOSes set different C-states for hyperthreads on the same
core, which prevents the core from entering deep C-states, since not
all hyperthreads of the core are in the deep C-states.

We fix this by using same number of C-states for hyperthreads on the
same core; the hyperthread w/ more C-states is taken as master.