kernel/acpi_thermal: If the _TMP can't be evaluated, increase the interval.
If _TMP evaluation fails two times, switch to trying every two hours.
I think eventually we should move to stop reading the sensor at all
(it is usually a bug in the AML and won't spring to life again all
of a sudden), but for now, this lowers the burden on log files.

Submitted-by: dillon