kernel/acpi: Reduce code duplication with ACPICA (second attempt).
Use the types which the ACPICA framework provides whereever possible in
our acpi_fadt and acpi_madt code. Since we do not use any actual ACPICA
functions here, this is no problem even for code which (so far) runs
before ACPICA is initialized.

Compared to the previous attempt which I pushed (and reverted again)
this commit has the following additional fixes:

* Compare the FADT length to the length of the ACPI 2.0 (up to 4.0)
  version of the table (which is 24 bytes shorter than the ACPI 5.0
  version).

* Doesn't shift again an already shifted value.

Discussed-with-and-approved-by: sephe
Tested-by: Robin Hahling <robin.hahling@gw-computing.net>