loader: set color option (blue) by default
On the loader screen, the blue version of Fred will now be shown by
default unless the serial console is selected ("comconsole" value for
"console" option in boot/loader.conf).

As part of the work:
  * "loader_color" option removed
  * "loader_plain" is a new option, to force Fred in monochrome
  * "fred_is_blue" option (4.1 only) removed
  * "fred_is_red"  is a new option as the red scheme is now alternate

Users that had "loader_color" set on before will see Fred change color to
blue, thus they will need to set "fred_is_red=true" in /boot/loader.conf
to restore the loader as it was before.  Users that don't like color will
have to explicitly set "loader_plain=true" to get rid of it.

This is a cosmetic change which hopefully makes DragonFly a tiny bit more
attractive to first time users.