cross-tools stage: Don't build unused linker (i.e. gold)
Only one linker is used to build the world, but each version of binutils
normally installs two linkers.  Update the build infrastructure to skip
building one linker for the cross-tools version of binutils.  This is
normally the gold linker, which builds very slowly due to being written
in c++.

This change could be backported to binutils 2.24, but right now it only
applies to binutils 2.25, the version normally used to build world.