binutils 2.24: Remove info pages
The info pages from binutils 2.25 are not installed (if they were, they
would overwrite 2.24 versions anyway).  Both "as" and "ld" have sufficient
man pages.  There is not equivalent "binutils" man page, but this
information is readily available on the internet, which is the same
justification used for not importing gcc50 info pages.