gcc50: Upgrade to GCC 5.1 release candidate plus
while here:
  * remove 4+ Mb of unused texi files from contrib
  * update the 3 man pages from Release candidate
  * Change makefiles to not take man pages from contrib directory

After thinking about it, I've decided to continue to update GCC from
Subversion sources rather than limit updates to designated releases.  This
is the reason the man pages are not taken from contrib; they are only
present in actual releases and not the SVN repository.  The plan is to
only update the man pages when GCC minor version is released.

This is the first (and likely only) release candidate for GCC5.  It will
be known as gcc 5.1.0, but CCVER will remain at "gcc50" regardless if the
version of gcc is 5.1.1 or 5.4.2.  This includes changes up to SVN 222168
(April 17).