hammer - Avoid panic in particular corruption situations
This commit fixes two panics which can arise from attempting to access a
corrupted filesystem.  It does not fix the issue that led to the corruption
in the first place (which we believe might have been GPU-related random
memory corruption on a particular machine).  It allows the filesystem to
continue to operate so it can be spooled off and reformatted.

* Do not panic on a NULL pointer dereference if a cursor iteration
  fails due to corruption.

* Do not panic during a flush on a sequence space error if the filesystem
  has been forced into read-only mode (which prevents older sequence spaces
  from being flushed).

Reported-by: Sephe