libc: Make generation of nslexer.c more robust.
Quoting FreeBSD's r268620 commit msg:

Ensure that lex errors fail the build instead of being silently ignored
due to the piped call.  Also postpone the update of the nslexer.c file
until we are sure we have generated it properly.

[...]

In DragonFly, this issue was causing a buildworld to stall if certain
-W... warning options were added.

Taken-from: FreeBSD